Extract shared ATA/mint helpers and the pool-state deserializer into reusable modules.
