+++ src/commands/spl-token/mint.ts
@@ -1,28 +1,13 @@
 import { Command } from 'commander';
 import { Connection, PublicKey, TransactionInstruction } from '@solana/web3.js';
-import { getAccount, ASSOCIATED_TOKEN_PROGRAM_ID } from '@solana/spl-token';
-import { detectTokenProgramId } from '../../utils/token.js';
+import { getAccount } from '@solana/spl-token';
+import { detectTokenProgramId, findAssociatedTokenAddress } from '../../utils/token.js';
 import { SplMintArgsSchema } from '../../types/index.js';
 import { validateArgs } from '../../utils/validation.js';
 import { TransactionBuilder } from '../../core/transaction-builder.js';
 import { finalizeTransaction } from '../../utils/finalize-transaction.js';
 import { logger } from '../../utils/logger.js';
 import { InstructionBuilder as SplInstructionBuilder } from '../../programs/spl-token/instructions.js';
-
-/**
- * Calculate ATA address without curve validation (for instruction building)
- */
-function findAssociatedTokenAddress(
-  mint: PublicKey,
-  owner: PublicKey,
-  tokenProgramId: PublicKey
-): PublicKey {
-  const [address] = PublicKey.findProgramAddressSync(
-    [owner.toBuffer(), tokenProgramId.toBuffer(), mint.toBuffer()],
-    ASSOCIATED_TOKEN_PROGRAM_ID
-  );
-  return address;
-}
 
 export async function mintCommand(options: Record<string, string>, command: Command) {
   try {
